requirements.txt
appdirs==1.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.2
filelock==3.0.12
imageio==2.9.0
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.3.56
pandas==1.3.1
pbr==5.6.0
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
rawutil==2.7.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
stevedore==3.3.0
threadpoolctl==2.1.0
tifffile==2021.7.2
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
